requirements.txt
discord.py
python-dotenv
pytz
huggingface_hub
requests
asyncio
diffusers
torch
transformers
accelerate
safetensors
